lua/events: report the scope a script runs in
Adds `trxc.events.is_level_script`, which tells a level script from a
global one. Modules of the public surface read it to end what a level
script made when the level ends, as its event handlers already do.
