+++ src/tests/lua/api/items.c
@@ -144,17 +144,6 @@ static void M_PushFake(lua_State *const L)
     lua_setfield(L, -2, "fire_kill");
 }
 
-// Event listeners attach against a script context; a global one lives for the
-// whole session, which is all these tests need.
-LUA_CONTEXT LUA_GetScriptContext(void)
-{
-    return LUA_CONTEXT_GLOBAL;
-}
-
-void LUA_SetScriptContext(const LUA_CONTEXT context)
-{
-}
-
 void ItemAction_SetInterceptor(const ITEM_ACTION_INTERCEPTOR interceptor)
 {
 }
