+++ src/trx/game/lua/capi/events.c
@@ -362,9 +362,19 @@ static int M_L_EventsDetach(lua_State *const L)
     return 1;
 }
 
+// trxc.events.is_level_script() -> bool
+static int M_L_EventsIsLevelScript(lua_State *const L)
+{
+    lua_pushboolean(L, LUA_GetScriptContext() == LUA_CONTEXT_LEVEL);
+    return 1;
+}
+
 static const luaL_Reg m_Module[] = {
-    { "attach", M_L_EventsAttach }, { "declare", M_L_EventsDeclare },
-    { "detach", M_L_EventsDetach }, { "fire", M_L_EventsFire },
+    { "attach", M_L_EventsAttach },
+    { "declare", M_L_EventsDeclare },
+    { "detach", M_L_EventsDetach },
+    { "fire", M_L_EventsFire },
+    { "is_level_script", M_L_EventsIsLevelScript },
     { nullptr, nullptr },
 };
